refactor: separate api module
