+++ js/components/TodoApp.js
@@ -2,7 +2,7 @@ import TodoUser from "./TodoUser.js";
 import TodoInput from "./TodoInput.js";
 import TodoList from "./TodoList.js";
 import TodoCount from "./TodoCount.js";
-import { User, Todo } from "../apis.js";
+import { User, Todo } from "../apis/index.js";
 
 export default function TodoApp(appEl) {
   this.init = async () => {
